+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Small tweaks to Aura of Protection scaling.
 * Monster: Nosferatu parsing tweaks to break bite damage out into versatile.
+* Monster: Dullahan had parsing problems around legendary/mythic actions and battle axe damage.
 * Rage could target another creature.
 * Some fixes to character import where if effects from things like active auras were present could prevent character re-imports working.